Further renaming The "dynatrace_client" package renamed to "dtclient", "oneagentutils" to "utils" and "oneagent/util.go" to "oneagent/oneagent_utils.go"
