+++ pkg/controller/nodes/controller_test.go
@@ -6,8 +6,8 @@ import (
 
 	apis "github.com/Dynatrace/dynatrace-oneagent-operator/pkg/apis"
 	dynatracev1alpha1 "github.com/Dynatrace/dynatrace-oneagent-operator/pkg/apis/dynatrace/v1alpha1"
-	oneagentutils "github.com/Dynatrace/dynatrace-oneagent-operator/pkg/controller/oneagentutils"
-	dtclient "github.com/Dynatrace/dynatrace-oneagent-operator/pkg/dynatraceclient"
+	"github.com/Dynatrace/dynatrace-oneagent-operator/pkg/controller/utils"
+	"github.com/Dynatrace/dynatrace-oneagent-operator/pkg/dtclient"
 	"github.com/operator-framework/operator-sdk/pkg/k8sutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -99,7 +99,7 @@ func TestNodesReconciler_NewSchedulableNode(t *testing.T) {
 
 	dtClient := &dtclient.MockDynatraceClient{}
 
-	nodesController := NewController(fakeClient, oneagentutils.StaticDynatraceClient(dtClient))
+	nodesController := NewController(fakeClient, utils.StaticDynatraceClient(dtClient))
 
 	assert.NoError(t, nodesController.ReconcileNodes(nodeName))
 	assert.Len(t, nodesController.nodeCordonedStatus, 1)
@@ -130,7 +130,7 @@ func TestNodesReconciler_UnschedulableNode(t *testing.T) {
 		return e.EventType == "MARKED_FOR_TERMINATION"
 	})).Return(nil)
 
-	nodesController := NewController(fakeClient, oneagentutils.StaticDynatraceClient(dtClient))
+	nodesController := NewController(fakeClient, utils.StaticDynatraceClient(dtClient))
 
 	assert.NoError(t, nodesController.ReconcileNodes(nodeName))
 	assert.Len(t, nodesController.nodeCordonedStatus, 1)
@@ -162,7 +162,7 @@ func TestNodesReconciler_DeletedNode(t *testing.T) {
 		return e.EventType == "MARKED_FOR_TERMINATION"
 	})).Return(nil)
 
-	nodesController := NewController(fakeClient, oneagentutils.StaticDynatraceClient(dtClient))
+	nodesController := NewController(fakeClient, utils.StaticDynatraceClient(dtClient))
 
 	assert.NoError(t, nodesController.ReconcileNodes(nodeName))
 	assert.Len(t, nodesController.nodeCordonedStatus, 1)
@@ -187,7 +187,7 @@ func TestNodesReconciler_UnschedulableNodeAndNoMatchingOneAgent(t *testing.T) {
 
 	dtClient := &dtclient.MockDynatraceClient{}
 
-	nodesController := NewController(fakeClient, oneagentutils.StaticDynatraceClient(dtClient))
+	nodesController := NewController(fakeClient, utils.StaticDynatraceClient(dtClient))
 
 	assert.NoError(t, nodesController.ReconcileNodes(nodeName))
 	assert.Empty(t, nodesController.nodeCordonedStatus)
