+++ pkg/controller/oneagent/oneagent_utils.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	dynatracev1alpha1 "github.com/Dynatrace/dynatrace-oneagent-operator/pkg/apis/dynatrace/v1alpha1"
-	dtclient "github.com/Dynatrace/dynatrace-oneagent-operator/pkg/dynatraceclient"
+	"github.com/Dynatrace/dynatrace-oneagent-operator/pkg/dtclient"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
