+++ pkg/dtclient/client_test.go
@@ -1,4 +1,4 @@
-package dynatraceclient
+package dtclient
 
 import (
 	"github.com/stretchr/testify/assert"
