+++ pkg/dtclient/communication_hosts_test.go
@@ -1,4 +1,4 @@
-package dynatraceclient
+package dtclient
 
 import (
 	"net/http"
