+++ pkg/dtclient/send_event.go
@@ -1,4 +1,4 @@
-package dynatraceclient
+package dtclient
 
 import (
 	"bytes"
